Quick hotfix to gaminglabAPI.py.
Getting the PC availability occurs in a loop. It will continue in 5 seconds if there happens to be a stale reference. The loop breaks if it successfully extracts the necessary web data. Having this loop is acceptable if the nature of the function is asynchronous- no other processes will be blocked, and it's absolutely critical to extract the necessary data before attempting to update the embed.

The following commit message is from a previous commit message, and still retains relevance in this aspect.

=====================================================

Hotfix to gaminglabAPI.py and new FAQ feature.

gaminglabAPI.py:

- The bot waits 5 seconds if there is initial failure to locate the website's iframes for webscraping. Immediate retries, as previously included, were insufficient in locating the necessary DOM elements.

bot.py:

- Added support for custom `!faq` command.

 gsheetsAPI.py:

 - Added support to retrieve the new FAQ sheet that was added to the Google Sheets config file. Responsible for pulling the necessary questions and answers.

 customcommands.py:

 - Added a new function `send_faq_panel` to iterate through the entered questions and answers as entered in the Google Sheets config file. The bot will notify the user if an FAQ channel is missing.

 gaminglabAPI.py:

 - Added hotfix for the frequent "stale reference element" exception caught by Selenium. As opposed to breaking from the loop, it's more optimal for the bot to retry in x amount of seconds. That is- as opposed to breaking from threaded function immediately and then attempting an automatic retry.