Temporarily disabled gaming lab feed. Will be replaced in the future by connecting to the gamelabs API. Had to temporarily relocated the gsheetsAPI.py module in order to actually receive the Google Sheets updates. It is currently in bot.py unused, but should be immediately relocated as soon as the means are available.
In gsheetsAPI.py, I was incorrectly accessing the authorized user ID values. The data type in question is a tuple, and I was grabbing the entire tuple set as opposed to the individual value. However, these values are hardcoded and a better solution should be looked into for a more dynamic setup.